unit test for cellno mapper
